refactor: util 패키지를 common으로 변경
